서보파라미터 수신 테스트중

+++ MINASA6SF_Rev/Models/ServoParaModel.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Documents;
 
 namespace MINASA6SF_Rev.Models
 {
@@ -75,8 +76,8 @@ namespace MINASA6SF_Rev.Models
         }
 
 
-        private double setval;
-        public double SetVal
+        private float setval;
+        public float SetVal
         {
             get
             {
